# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "RPN Calculator API",
        "version": "1.0.0"
    },
    "paths": {
        "/rpn/stack": {
            "get": {
                "tags": [
                    "RPN"
                ],
                "summary": "List Stacks",
                "operationId": "list_stacks_rpn_stack_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "RPN"
                ],
                "summary": "Create Stack",
                "operationId": "create_stack_rpn_stack_post",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/rpn/stack/{stack_id}": {
            "get": {
                "tags": [
                    "RPN"
                ],
                "summary": "Get Stack",
                "operationId": "get_stack_rpn_stack__stack_id__get",
                "parameters": [
                    {
                        "name": "stack_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "title": "Stack Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "RPN"
                ],
                "summary": "Push Value",
                "operationId": "push_value_rpn_stack__stack_id__post",
                "parameters": [
                    {
                        "name": "stack_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "title": "Stack Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PushRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "RPN"
                ],
                "summary": "Delete Stack",
                "operationId": "delete_stack_rpn_stack__stack_id__delete",
                "parameters": [
                    {
                        "name": "stack_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "title": "Stack Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/rpn/op": {
            "get": {
                "tags": [
                    "RPN"
                ],
                "summary": "List All Operands",
                "operationId": "list_all_operands_rpn_op_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/rpn/op/{op}/stack/{stack_id}": {
            "post": {
                "tags": [
                    "RPN"
                ],
                "summary": "Apply Operation",
                "operationId": "apply_operation_rpn_op__op__stack__stack_id__post",
                "parameters": [
                    {
                        "name": "op",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Operation"
                        }
                    },
                    {
                        "name": "stack_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "title": "Stack Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "Operation": {
                "type": "string",
                "enum": [
                    "add",
                    "sub",
                    "mul",
                    "div"
                ],
                "title": "Operation"
            },
            "PushRequest": {
                "properties": {
                    "value": {
                        "type": "number",
                        "title": "Value"
                    }
                },
                "type": "object",
                "required": [
                    "value"
                ],
                "title": "PushRequest"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}